refactor: tests to improve consistency and readability
- Updated various test classes to use `await using` for disposable resources.
- Added `AbortToken` parameter to several asynchronous factory methods for better cancellation support.
- Removed unnecessary using directives to clean up code.
- Simplified constructor initialization for test classes.
- Enhanced exception handling in tests to ensure clarity and maintainability.
- Adjusted assertions to use specific exception types for better accuracy.

+++ tests/Headless.Checks.Tests.Unit/IsNegativeOrZeroTests.cs
@@ -91,7 +91,7 @@ public sealed class IsNegativeOrZeroTests
         {
             NumericKind.Short => () =>
             {
-                const short value = (short)3;
+                const short value = 3;
                 Argument.IsNegativeOrZero(value, message);
             },
             NumericKind.Int => () =>
